Support downstream middleware with npm dependencies
