--- server/is-fastboot-route.js
@@ -1,3 +1,0 @@
-const isFastBootRoute = (path) => true;
-
-module.exports = isFastBootRoute;
