+++ server/middleware/is-fastboot-route.js
@@ -1,0 +1,3 @@
+const isFastbootRoute = (path) => true;
+
+module.exports = isFastbootRoute;
