+++ server/middleware/post-fastboot.js
@@ -1,0 +1,4 @@
+"use strict";
+
+module.exports = [
+];
